+++ src/main/java/com/zea7ot/leetcode/tags/data_structure/map/hashmap.txt
@@ -9,6 +9,7 @@ References:
 1   0884    https://leetcode.com/problems/uncommon-words-from-two-sentences/
 
 2   0001    https://leetcode.com/problems/two-sum/
+2   0170    https://leetcode.com/problems/two-sum-iii-data-structure-design/
 2   0290    https://leetcode.com/problems/word-pattern/
 2   0791    https://leetcode.com/problems/custom-sort-string/
 2   0953    https://leetcode.com/problems/verifying-an-alien-dictionary/
