+++ src/main/java/com/zea7ot/leetcode/tags/data_structure/stream/summary.txt
@@ -3,7 +3,9 @@ References:
     https://github.com/huanminwu/LeetCode/blob/master/LeetCode.xlsx
 
 
-### Array
-# Stream
+### Stream
+# Summary
 
 1   0346    https://leetcode.com/problems/moving-average-from-data-stream/
+
+2   0170    https://leetcode.com/problems/two-sum-iii-data-structure-design/
